backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
anthropic>=0.25.0
openai>=1.12.0
pydantic>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
cryptography>=42.0.0
psycopg2-binary>=2.9.0
alembic>=1.13.0
